requirements.txt
amqp==5.2.0
APScheduler==3.10.4
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron-descriptor==1.4.3
cryptography==42.0.7
daphne==4.1.2
Django==4.2.13
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==24.2
django-phonenumber-field==7.3.0
django-timezone-field==6.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
drf-spectacular==0.27.2
drf-yasg==1.21.7
et-xmlfile==1.1.0
eventlet==0.36.1
flower==2.0.1
gevent==24.2.1
greenlet==3.0.3
h11==0.14.0
humanize==4.9.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kombu==5.3.7
msgpack==1.0.8
numpy==1.26.4
openpyxl==3.1.3
packaging==24.0
pandas==2.2.2
phonenumbers==8.13.36
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
referencing==0.35.1
requests==2.31.0
rpds-py==0.18.1
service-identity==24.1.0
six==1.16.0
SQLAlchemy==2.0.30
sqlparse==0.5.0
tornado==6.4
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.29.0
vine==5.1.0
wcwidth==0.2.13
zope.event==5.0
zope.interface==6.4
